ZEPPELIN-3983. Travis fails due to downloading spark takes a lot of time
